staging: vcsm-cma: Rework to use dma APIs, not CMA
Due to a misunderstanding of the DMA mapping APIs, I made
the wrong decision on how to implement this.

Rework to use dma_alloc_coherent instead of the CMA
API. This also allows it to be built as a module easily.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
(cherry picked from commit a9d1cec3451c379bcbc37ee47ffe31f58792907a)
Signed-off-by: Hui Wang <hui.wang@canonical.com>