UBUNTU: SAUCE: add one more flavour to armhf kernel
https://bugs.launchpad.net/bugs/1853446

Add one more flavour raspi4 to the armhf kernel, after that the armhf
kernel has two flavours, the flavour raspi2 disable the LPAE, HIGHMEM
and VMSPLIT_3G, so this kernel is only for rpi2/3 boards, the other
flavour raspi4 enable the LPAE, HIGHMEM and VMSPLIT_3G, it is for
rpi4 boards. But in theory the flavour raspi4 could work on rpi2/3/4
boards.

Signed-off-by: Hui Wang <hui.wang@canonical.com>